refactor: created cli and gui packages inside infrastructure package
