+++ src/infrastructure/gui/ErrorCode.py
