+++ src/infrastructure/gui/presentation/TitleCollection.py
